Add `exchange` parameter to AMQP External Rule
This commit adds an `exchange` parameter to the [ably_rule_amqp_external Resource](https://registry.terraform.io/providers/ably/ably/latest/docs/resources/rule_amqp_external).

The `exchange` parameter allows specifying a RabbitMQ exchange. If required, this parameter supports interpolation.
See [Ably FAQ](https://faqs.ably.com/what-is-the-format-of-the-routingkey-for-an-amqp-or-kinesis-reactor-rule) for information.
If RabbitMQ exchanges are not used, this parameter should not be specified in the Terraform resource.

Addresses https://github.com/ably/terraform-provider-ably/issues/159