Use Terraform types package
Since we're redefining all our own types for the Terraform provider, we
may as use the Terraform `types` package, which provides additional
safety.

If we're not using these native types, then why do we bother redefining
the types? Perhaps the answer is that we don't need to, but this feels
like the correct thing to do.