Fix bug with inconsistent return of tlsOnly
When testing the fcmServiceAccount, I did not specify `tls_only`, and it
produced inconsistent results because a missing value was defaulting to
`false`.

Ideally we should use pointer values because the Control API should be
able to handle `null` values, and using pointers would simplify this.

For now, we will just default it to `false`.